Updated Comments for intellisense
I fixed many of the comments that were incorrectly wrapped (The summary
tag does not wrap around the param tags).  Now, at least for the classes
I've updated the intellisense should have a bit more data.  I added some
of the intellisense information from their past API docs.  As I clarify
the changes in this incarnation of their service, I'll correct and
update the rest of the references.